test: try to de-flake
